Created a cleaner directory structure Better header file (as far as I know) Put structure definitions into header file Added Makefile (Edited version of the one in Video-Editing-Automation) to be able to build from source in Linux environments Added Linux build instructions to README.md
