+++ Samples/maze.txt
@@ -1,0 +1,9 @@
+#S###############
+#OOO#O#OOOOO#OOO#
+#O#O#O#O###O###O#
+#O#OOO#O#O#O#OOO#
+#O#######O#O#O###
+#OOO#OOO#OOO#OOO#
+#O#O#O#O#O#####O#
+#O#OOO#OOOOOOOOOG
+#################
