+++ Samples/sampleMaze2.txt
@@ -1,0 +1,21 @@
+###S#################
+#O#OOOOOOOOO#OOOOO#O#
+#O#O###O#####O#O#O#O#
+#OOO#O#OOOOO#O#O#OOO#
+#O###O#####O#O#O###O#
+#O#OOOOO#OOO#O#O#OOO#
+#O###O#O#O###O#O#####
+#OOO#O#O#O#OOO#OOOOO#
+###O###O#O#O#####O#O#
+#O#OOO#O#O#O#OOO#O#O#
+#O###O#O#O#O###O#O#O#
+#OOO#OOO#OOOOO#O#O#O#
+#O#####O#######O###O#
+#OOO#OOO#OOO#OOO#OOO#
+#O#O#O###O#O#O###O###
+#O#OOO#OOO#O#O#OOOOO#
+#O#######O###O#####O#
+#OOO#OOOOO#OOO#OOO#O#
+###O#O###O#O###O#O#O#
+#OOOOO#OOO#OOOOO#OOO#
+###############G#####
