+++ include/funcDefns.h
@@ -1,3 +1,27 @@
+#ifndef funcDefns_h
+#define funcDefns_h
+
+// Creating GridSize struct
+typedef struct GridSize {
+	int width;
+	int height;
+} GridSize;
+
+// Creating Node struct
+typedef struct Node {
+	struct Node* parent;
+	int exists;
+	int x;
+	int y;
+	int g;
+	int h;
+	int f;
+	int id;
+} Node;
+
+GridSize size;
+GridSize getGridSize(FILE *file);
+
 int readFile(FILE *file, char grid[][size.width]);
 void initString(char string[]);
 void appendName(char origName[], char insert[], char newName[]);
@@ -24,3 +48,5 @@ void insertPath(char grid[][size.width], Node path[][size.width]);
 void info(Node node);
 void info1D(Node nodes[], int length);
 void info2D(Node nodes[][size.width], int length);
+
+#endif
